Move test helpers to their own module
