+++ frontend/src/utils/validate/helpers.ts
@@ -1,10 +1,9 @@
 /* eslint-disable no-param-reassign */
 
-import { defaultsDeep, get, hasIn, set } from 'lodash';
-import { DeepPartial, PickByValue } from 'utility-types';
+import { get, set } from 'lodash';
+import { PickByValue } from 'utility-types';
 import { z } from 'zod';
 
-import { DEFAULT_PLUGIN_DATA } from '@/constants/plugin';
 import { PluginData, PluginIndexData } from '@/types';
 
 /**
@@ -102,54 +101,3 @@ export function getStringArraySanitizer<T extends PluginIndexData | PluginData>(
  * Type for declaring a map of functions that are used to sanitize a particular object prop.
  */
 export type SanitizerMap<T> = Partial<Record<keyof T, (result: T) => void>>;
-
-function getPluginData(data: DeepPartial<PluginData>) {
-  return defaultsDeep(data, DEFAULT_PLUGIN_DATA) as PluginData;
-}
-
-/**
- * Symbol for testing props that should be deleted.
- */
-export const DELETED = Symbol('deleted');
-
-/**
- * Creates new test cases for validating plugin data against valid and invalid data.
- */
-export function createValidatePluginTest<T, K extends keyof T = keyof T>({
-  key,
-  validData,
-  invalidData,
-  expectedInvalidDataResult,
-  validatePlugin,
-}: {
-  key: K;
-  validData: DeepPartial<T[K]>;
-  invalidData: DeepPartial<T[K]>;
-  expectedInvalidDataResult: DeepPartial<T[K]> | typeof DELETED;
-  validatePlugin: (plugin: T) => T;
-}) {
-  // eslint-disable-next-line jest/valid-title
-  describe(key as string, () => {
-    it('should return data if valid', () => {
-      const plugin = getPluginData({
-        [key]: validData,
-      });
-
-      const result = validatePlugin(plugin as unknown as T);
-      expect(result[key]).toEqual(get(plugin, key));
-    });
-
-    it('should return defaults if invalid', () => {
-      const plugin = getPluginData({
-        [key]: invalidData,
-      });
-
-      const result = validatePlugin(plugin as unknown as T);
-      if (expectedInvalidDataResult === DELETED) {
-        expect(hasIn(result, key)).toBe(false);
-      } else {
-        expect(result[key]).toEqual(expectedInvalidDataResult);
-      }
-    });
-  });
-}
